kivy_app/requirements.txt
kivy
Pillow
python-barcode
arabic-reshaper
python-bidi
